requirements.txt
ads==0.12.3
certifi==2021.10.8
charset-normalizer==2.0.12
httpretty==0.8.10
idna==3.3
mock==4.0.3
pbr==5.8.1
requests==2.27.1
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.3
XlsxWriter==3.0.2

python-dotenv==0.17.1
numpy==1.22.2
pandas==1.4.1
weasyprint==54.1